Do not force remove data directories if they are not empty.
pgxc_ctl would try to cleanup specified directories (using rm -rf) even if they
are not empty. It does so for GTM, datanodes, coordinators and all other
components. This could lead to serious issues if user accidently specifies a
wrong directory since it will wipe out everything under that directory.

We now check if the directory is empty and then only proceed. Otherwise the
initialization step will be skipped and existing data directory will be used to
start the corresponding server. Also, we do this only for critical components
such as datanodes, coordinators and GTM master. Other components such as
datanode/coordinator/GTM slaves and GTM proxies can be recreated without any
data loss.

+++ contrib/pgxc_ctl/coord_cmd.c
@@ -77,6 +77,14 @@ cmd_t *prepare_initCoordinatorMaster(char *nodeName)
 			 nodeName);
 		return(NULL);
 	}
+
+	if (pgxc_check_dir(aval(VAR_coordMasterDirs)[jj]) == 2)
+	{
+		elog(ERROR, "ERROR: target coordinator directory %s exists and is not empty. Skip initilialization.\n",
+				aval(VAR_coordMasterDirs)[jj]); 
+		return NULL;
+	}
+
 	cmd = cmdInitdb = initCmd(aval(VAR_coordMasterServers)[jj]);
 	snprintf(newCommand(cmdInitdb), MAXLINE, 
 			 "rm -rf %s;"
